added permissions for stocks

+++ routes/api/materials.php
@@ -5,18 +5,18 @@ use App\Http\Controllers\MaterialStoreController;
 use Illuminate\Support\Facades\Route;
 
 Route::prefix('materials')->middleware('auth:sanctum')->controller(MaterialController::class)->group(function () {
-  Route::get('/', 'index');
-  Route::get('/{id}', 'show');
-  Route::post('/', 'store');
-  Route::put('/{id}', 'update');
-  Route::delete('/{id}', 'destroy');
+  Route::get('/', 'index')->middleware('permission:stockMaterial_list');
+  Route::get('/{id}', 'show')->middleware('permission:stockMaterial_display');
+  Route::post('/', 'store')->middleware('permission:stockMaterial_insert');
+  Route::put('/{id}', 'update')->middleware('permission:stockMaterial_update');
+  Route::delete('/{id}', 'destroy')->middleware('permission:stockMaterial_delete');
 });
 
 Route::prefix('material_store')->middleware('auth:sanctum')->controller(MaterialStoreController::class)->group(function () {
-  Route::get('/', 'index');
-  Route::post('/', 'store');
-  Route::get('/{id}', 'show');
-  Route::put('/{id}', 'update');
-  Route::delete('/{id}', 'destroy');
-  Route::patch('/{id}/limits', 'updateLimits');
+  Route::get('/', 'index')->middleware('permission:stockMaterial_list');
+  Route::post('/', 'store')->middleware('permission:stockMaterial_insert');
+  Route::get('/{id}', 'show')->middleware('permission:stockMaterial_display');
+  Route::put('/{id}', 'update')->middleware('permission:stockMaterial_update');
+  Route::delete('/{id}', 'destroy')->middleware('permission:stockMaterial_delete');
+  Route::patch('/{id}/limits', 'updateLimits')->middleware('permission:stockMaterial_update');
 });
